+++ lib/braintree/three_d_secure_info.rb
@@ -2,16 +2,21 @@ module Braintree
   class ThreeDSecureInfo # :nodoc:
     include BaseModule
 
-    attr_reader :enrolled
-    attr_reader :liability_shifted
-    attr_reader :liability_shift_possible
-    attr_reader :status
+    attr_reader :acs_transaction_id
     attr_reader :cavv
-    attr_reader :xid
-    attr_reader :eci_flag
-    attr_reader :three_d_secure_version
     attr_reader :ds_transaction_id
+    attr_reader :eci_flag
+    attr_reader :enrolled
+    attr_reader :liability_shift_possible
+    attr_reader :liability_shifted
+    attr_reader :pares_status
+    attr_reader :status
     attr_reader :three_d_secure_authentication_id
+    attr_reader :three_d_secure_transaction_id
+    attr_reader :three_d_secure_version
+    attr_reader :xid
+    attr_reader :lookup
+    attr_reader :authentication
 
     alias_method :liability_shifted?, :liability_shifted
     alias_method :liability_shift_possible?, :liability_shift_possible
@@ -22,16 +27,21 @@ module Braintree
 
     def inspect
       attr_order = [
-        :enrolled,
-        :liability_shifted,
-        :liability_shift_possible,
-        :status,
+        :acs_transaction_id,
+        :authentication,
         :cavv,
-        :xid,
-        :eci_flag,
-        :three_d_secure_version,
         :ds_transaction_id,
+        :eci_flag,
+        :enrolled,
+        :liability_shift_possible,
+        :liability_shifted,
+        :lookup,
+        :pares_status,
+        :status,
         :three_d_secure_authentication_id,
+        :three_d_secure_transaction_id,
+        :three_d_secure_version,
+        :xid
       ]
 
       formatted_attrs = attr_order.map do |attr|