+++ spec/unit/braintree/risk_data_spec.rb
@@ -2,12 +2,14 @@ require File.expand_path(File.dirname(__FILE__) + "/../../spec_helper")
 
 describe Braintree::RiskData do
   describe "#initialize" do
-    it "sets id, decision and device_data_captured" do
-      risk_data = Braintree::RiskData.new(:id => "123", :decision => "YOU WON $1000 DOLLARS", :device_data_captured => true, :fraud_service_provider => "kount")
+    it "sets id, decision, device_data_captured, decision_reasons and transaction_risk_score" do
+      risk_data = Braintree::RiskData.new(:id => "123", :decision => "YOU WON $1000 DOLLARS", :device_data_captured => true, :fraud_service_provider => "fraud_protection", :decision_reasons => ["reason"], :transaction_risk_score => "12")
       risk_data.id.should == "123"
       risk_data.decision.should == "YOU WON $1000 DOLLARS"
       risk_data.device_data_captured.should be_truthy
-      risk_data.fraud_service_provider.should == "kount"
+      risk_data.fraud_service_provider.should == "fraud_protection"
+      risk_data.decision_reasons.should == ["reason"]
+      risk_data.transaction_risk_score.should == "12"
     end
   end
 
@@ -16,10 +18,12 @@ describe Braintree::RiskData do
       details = Braintree::RiskData.new(
         :id => "123",
         :decision => "YOU WON $1000 DOLLARS",
+        :decision_reasons => ["reason"],
         :device_data_captured => true,
-        :fraud_service_provider => "kount",
+        :fraud_service_provider => "fraud_protection",
+        :transaction_risk_score => "12",
       )
-      details.inspect.should == %(#<RiskData id: "123", decision: "YOU WON $1000 DOLLARS", device_data_captured: true, fraud_service_provider: "kount">)
+      details.inspect.should == %(#<RiskData id: "123", decision: "YOU WON $1000 DOLLARS", decision_reasons: ["reason"], device_data_captured: true, fraud_service_provider: "fraud_protection", transaction_risk_score: "12">)
     end
   end
 end