+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 3.3.0
+* Add `decision_reasons` and `transaction_risk_score` fields to `RiskData`
+* Add `acs_transaction_id`, `pares_status`, `three_d_secure_transaction_id`, `lookup`, and `authentication` to `ThreeDSecureInfo`
+
 ## 3.2.0
 * Add SCA exemption to Transaction sale
 * Add Installment support to Transaction sale